rename: config, helper 디렉토리 분리
- config 디렉토리 : 해당 어플리케이션에서 사용되는 설정 및 상수 관리 (설정, 메세지)
- helper 디렉토리 : 어플리케이션과 성격과 관계없이 독립적으로 사용되는 함수 (로컬스토리지 핸들링)
reference by https://codeigniter4.github.io/CodeIgniter4/concepts/structure.html
