+++ src/ts/config/configuration.ts
